Fix NoneType error in describe_images by filtering out None values in… (#1310)
This merge request fixes an issue where the `describe_images` function
in `cartography/intel/aws/ec2/images.py` encountered a NoneType error.
The error occurred because None values were being included in the
image_ids list, which is not allowed by the AWS SDK. In one particular
account of several hundred tested, an image ID was returned that was
None. I was able to verify that this allows all of my accounts to run
ingestion successfully.