Fix #1334: add retry + sleep to github team-repo sync (#1336)
### Summary
> Describe your changes.

Adds a retry with sleep to `_get_team_repos_for_multiple_teams` because
we have seen the GitHub GraphQL API sometimes return None for fields
that are not supposed to be None.

### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/1334

### Proof that this works
> We can merge your change in faster if we see that it works. For
example, if making a change to the graph, include a
> screenshot showing what the graph looked like before and after your
changes. You can also include console log traces
> showing what happened before and after your changes.

Added unit tests.

### Checklist

- [x] Update/add unit or integration tests

If you are modifying or implementing an intel module:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md)
- [ ] Use the NodeSchema [data
model](https://lyft.github.io/cartography/dev/writing-intel-modules.html#defining-a-node)

Signed-off-by: chandanchowdhury <chandan.chowdhury@hotmail.com>