Add support for fuzzy case-insensitive match in PropertyRef (#1383)
### Summary
> Describe your changes.

We often need to connect 2 nodes based on a fuzzy match like in
https://github.com/cartography-cncf/cartography/pull/1380#discussion_r1837801778.
This PR adds a `fuzzy_and_ignore_case` option to the `PropertyRef` so
that the final rendered query performs a `CONTAINS` in a
case-insensitive way instead of an exact match.

### Related issues or links
> Include links to relevant issues or other pages.

https://github.com/cartography-cncf/cartography/pull/1380

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

---------

Signed-off-by: Alex Chantavy <achantavy@lyft.com>
Signed-off-by: chandanchowdhury <chandan.chowdhury@hotmail.com>