Auto scaling groups data model refactor (#1394)
### Summary

Migrating the ASG sync to use the data model. Nothing else was changed
in the sync aside from creating the model and using it during the sync.

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Sergio Franco <sergiof@lyft.com>
Signed-off-by: cmm-lyft <181004184+cmm-lyft@users.noreply.github.com>