CI: Wait for PyPI to publish before building Dockerfile (#1437)
### Summary
> Describe your changes.

When making a new release, we
1. publish the release to PyPI
2. install it from PyPI to the Dockerfile
3. publish that Dockerfile to GHCR

After publishing to PyPI however we need a few more seconds for the
package to be available. Currently we skip over this and the latest
release is not available, causing the build to fail.

![image](https://github.com/user-attachments/assets/4d7aa967-d0e6-4ed8-8ef1-be25b573eb06)

https://github.com/cartography-cncf/cartography/actions/runs/12577359450/job/35054696773

This PR makes it so that we now wait 10 seconds for PyPI before
attempting to build the image.

Signed-off-by: Alex Chantavy <chantavy@gmail.com>
Signed-off-by: Eryx Paredes <eryxp@lyft.com>