feat: Add S3AccountPublicAccessBlock support to AWS module (#1570)
### Summary
I have added S3AccountPublicAccessBlock Support to the AWS module. I am
working on adding the remaining intel modules and prioritised this
specific one because I feel this is important from the security point of
view.

### Related issues or links

- https://github.com/cartography-cncf/cartography/issues/1477

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [x] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: AdityaPandeyCN <adityapand3y666@gmail.com>
Signed-off-by: Jeremy Chapeau <jeremy.chapeau@gmail.com>