feat: #1625 Add object_ownership property to S3 Buckets (#1626)
### Summary
This PR adds the [object
ownership](https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html)
S3 bucket setting as a property on S3Bucket nodes. It calls the
[GetBucketOwnershipControls](https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketOwnershipControls.html)
API action, which currently returns a list with exactly one entry, the
bucket's object ownership rule. The value of that rule is added to the
S3Bucket node as the `object_ownership` property. I'm open to renaming
the property if you think it's confusing.
If the Cartography job doesn't have permission to call
`GetBucketOwnershipControls`, the API returns
`OwnershipControlsNotFoundError`, which this logs as a warning.

See [https://github.com/cartography-cncf/cartography/issues/1625] for
why I think this property is useful.

### Related issues or links
- https://github.com/cartography-cncf/cartography/issues/1625

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

![cartography_s3_properties](https://github.com/user-attachments/assets/82bc22c6-445d-43e7-b8e8-5bbdcc808b4b)
]

---------

Signed-off-by: Justin Owen <jowen@foursquare.com>
Signed-off-by: Jeremy Chapeau <jeremy.chapeau@gmail.com>