Refactor AWS Inspector: reduce memory and add parallel ingestion (#1654)
### Summary
There is a memory issue with AWS Inspector module. The pagination only
returns results until findings are retrieved from one account and all
those objects are added to one single list. Considering there can be
millions of findings, this can consume all the memory of the job running
it.
This refactor:
- Provide a maximum number of pages to `aws_paginate`
- Transform `aws_paginate` and `batch` into generators.
- Breaks inspector findings ingestion in batches of 1000 elements.
- Adds aysnc processing to ingest multiple accounts in parallel

---------

Signed-off-by: Eryx Paredes <eryxp@lyft.com>
Signed-off-by: Jeremy Chapeau <jeremy.chapeau@gmail.com>