Revert "feat: EntraID best effort mode" (#1774)
Reverts the Entra best-effort-mode in cartography-cncf/cartography#1751.


## Rationale

After some reflection and discussion, it doesn't make sense to have a
best-effort-mode for Entra at this time. Best effort mode was a hack
created for AWS because there are so many resources in AWS that can fail
in unpredictable ways. Entra has much fewer resources.

Our original plan to do this for Entra was misguided: we wanted to catch
Entra resources that were not configured properly and then continue on
to those that were. After discussion, it'd be a better approach here to
catch on the specific exception and handle things in a more tailored way
by re-raising with appropriate context. We'd rather let the aws module
be an outlier with best effort mode existing only there.