fix: #1591 Analysis job failure: aws_s3acl_analysis.json (#1795)
### Summary

Fix incorrect location of the `aws_s3acl_analysis.json` analysis job
file, which caused errors when running scoped analysis.

### Context

The root cause of this bug was that the `aws_s3acl_analysis.json` job,
designed to run with a `scoped_analysis` parameter, was not placed in
the correct data folder.
When using the folder directly, this mismatch resulted in a failure to
locate the analysis job file.

This analysis is explicitly invoked in the AWS module. The code has been
updated to call the correct file.

### Related issues or links

- #1591

### Changes proposed

* **Move** `aws_s3acl_analysis.json` to the correct scoped analysis
directory.
* **Update** AWS module to reference the correct file path.

Signed-off-by: Jeremy Chapeau <jeremy.chapeau@gmail.com>
Signed-off-by: wahidullah <wahidmaqs49@gmail.com>